chore(booking service):create booking service that does the business logic of the app
[Delivers #167750530]
